# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive API for managing prompts, snippets, notes, and templates for AI interactions.\n\n## Features\n- **Prompts**: Create, manage, and organize prompts for various AI models\n- **Snippets**: Store and manage reusable code snippets and text blocks\n- **Notes**: Add contextual notes to prompts for better organization\n- **Templates**: Preview and analyze template content\n- **Tagging**: Organize content with flexible tagging system\n- **Linking**: Create relationships between prompts",
        "title": "Proompt API",
        "contact": {
            "name": "Proompt API Support",
            "url": "https://github.com/dikkadev/proompt",
            "email": "support@proompt.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/template/analyze": {
            "post": {
                "description": "Analyze a template to extract variables, functions, and structure information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Analyze template structure",
                "parameters": [
                    {
                        "description": "Template analysis data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TemplatePreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template analysis result",
                        "schema": {
                            "$ref": "#/definitions/models.TemplatePreviewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or template syntax",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/template/preview": {
            "post": {
                "description": "Preview how a template will render with provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Preview template rendering",
                "parameters": [
                    {
                        "description": "Template preview data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TemplatePreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template preview result",
                        "schema": {
                            "$ref": "#/definitions/models.TemplatePreviewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or template syntax",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.TemplatePreviewRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "variables": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "models.TemplatePreviewResponse": {
            "type": "object",
            "properties": {
                "resolved_content": {
                    "type": "string"
                },
                "variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TemplateVariable"
                    }
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.TemplateVariable": {
            "type": "object",
            "properties": {
                "default_value": {
                    "type": "string"
                },
                "has_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "description": "\"provided\", \"default\", \"missing\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for authentication (if implemented)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}